backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
python-decouple==3.8
Pillow==10.1.0
